+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ SAP Operations Collection Release Notes
 .. contents:: Topics
 
 
-v1.12.0
+v1.13.0
 =======
 
 Release Summary
@@ -13,10 +13,26 @@ Release Summary
 
 Feature release
 
+Major Changes
+-------------
+
+- Add ABAP transport management modules
+
 New Modules
 -----------
 
-- sap.sap_operations.hana_restore - Restore SAP HANA database backup
+- sap.sap_operations.abap_transport_create - Create transport request in SAP ABAP system
+- sap.sap_operations.abap_transport_import - Import transport request
+- sap.sap_operations.abap_transport_info - Fetch information about transport request from SAP ABAP system
+- sap.sap_operations.abap_transports_info - Fetch transport requests from SAP ABAP system based on search criteria
+
+v1.12.0
+=======
+
+Release Summary
+---------------
+
+Feature release
 
 v1.11.0
 =======